Allow use of io.Readers for Doc and Value in iterators
This won't make much difference for normal CouchDB usage, where each JSON
segment must be parsed in turn.

The immediate reason for this is to allow consolidating the Row and Rows
type for the purpose of https://github.com/go-kivik/kivik/issues/398

This makes it unnecessary document into memory, in the case of a simple
single-document Get(), once these types are consolidated.

This may also allow optimizations for other backends in the future, to not
load an entire document into memory before needed.

This also retains full backward compatibility, by duplicating the old
interface to `driver.OldRow`.  This can be eliminated after a resonable
period.